add text splitter
