+++ src/features/chat/chat-services/chat-api-data.ts
@@ -2,9 +2,9 @@ import { userHashedId } from "@/features/auth/helpers";
 import { OpenAIInstance } from "@/features/common/openai";
 import { AI_NAME } from "@/features/theme/customise";
 import { OpenAIStream, StreamingTextResponse } from "ai";
-import { CosmosDBChatMessageHistory } from "../memory/cosmosdb/cosmosdb";
-import { similaritySearchVectorWithScore } from "../vector-stores/azure-cog-search/azure-cog-vector-store";
+import { similaritySearchVectorWithScore } from "./azure-cog-search/azure-cog-vector-store";
 import { initAndGuardChatSession } from "./chat-thread-service";
+import { CosmosDBChatMessageHistory } from "./cosmosdb/cosmosdb";
 import { PromptGPTProps } from "./models";
 
 const SYSTEM_PROMPT = `You are ${AI_NAME} who is a helpful AI Assistant.`;
