+++ src/features/chat/chat-services/chat-api-simple.ts
@@ -2,8 +2,8 @@ import { userHashedId } from "@/features/auth/helpers";
 import { OpenAIInstance } from "@/features/common/openai";
 import { AI_NAME } from "@/features/theme/customise";
 import { OpenAIStream, StreamingTextResponse } from "ai";
-import { CosmosDBChatMessageHistory } from "../memory/cosmosdb/cosmosdb";
 import { initAndGuardChatSession } from "./chat-thread-service";
+import { CosmosDBChatMessageHistory } from "./cosmosdb/cosmosdb";
 import { PromptGPTProps } from "./models";
 
 export const ChatAPISimple = async (props: PromptGPTProps) => {
